Redo

+++ ChinookSolution/ChinookSystem/BLL/AlbumController.cs
@@ -5,11 +5,12 @@ using System.Text;
 using System.Threading.Tasks;
 
 #region Additional Namespaces
-using ChinookSystem.Data.Entities;
-using ChinookSystem.DAL;
-using System.ComponentModel; //ODS
-using DMIT2018Common.UserControls;  //used by error handle user control
 using ChinookSystem.Data.POCOs;
+using ChinookSystem.DAL;
+using ChinookSystem.Data.Entities;
+using System.ComponentModel;
+using DMIT2018Common.UserControls;
+using ChinookSystem.Data.DTOs;
 #endregion
 
 namespace ChinookSystem.BLL
@@ -17,26 +18,21 @@ namespace ChinookSystem.BLL
     [DataObject]
     public class AlbumController
     {
-        //private data member to use with error handle messages
+        #region Class Variables
         private List<string> reasons = new List<string>();
+        #endregion
 
         #region Queries
         [DataObjectMethod(DataObjectMethodType.Select,false)]
-        //basic query: complete list of DbSet
         public List<Album> Album_List()
         {
-            //set up the code block to ensure the release of the sql connection
             using (var context = new ChinookContext())
             {
-                //.ToList<T> is used to convert the DbSet<T> into a List<T> collection
                 return context.Albums.ToList();
             }
-
         }
 
-        [DataObjectMethod(DataObjectMethodType.Select, false)]
-        //basic query: return a recorded based on pkey
-        public Album Album_FindByID(int albumid)
+        public Album Album_Get(int albumid)
         {
             using (var context = new ChinookContext())
             {
@@ -49,58 +45,91 @@ namespace ChinookSystem.BLL
         {
             using (var context = new ChinookContext())
             {
-                var results = from albumrow in context.Albums
-                              where albumrow.ArtistId == artistid
-                              select albumrow;
-                return results.ToList();
-            }
-        }
-
-        [DataObjectMethod(DataObjectMethodType.Select, false)]
-        public List<Album> Album_FindByTitle(string title)
-        {
-            using (var context = new ChinookContext())
-            {
-                var results = from albumrow in context.Albums
-                              where albumrow.Title.Contains(title)
-                              select albumrow;
-                return results.ToList();
-            }
-        }
-
-        [DataObjectMethod(DataObjectMethodType.Select, false)]
-        public List<AlbumArtists> Album_AlbumArtists()
-        {
-            using (var context = new ChinookContext())
-            {
                 var results = from x in context.Albums
-                              select new AlbumArtists
+                              where x.ArtistId == artistid
+                              select x;
+
+                return results.ToList();
+            }
+        }
+
+        [DataObjectMethod(DataObjectMethodType.Select,false)]
+        public List<AlbumsOfArtist> Album_AlbumsOfArtist(string artistname)
+        {
+            using (var context = new ChinookContext())
+            {
+                //unlike LinqPad which is Linq to Sql
+                //within out application it is Linq to Entities
+                var results = from x in context.Albums
+                              where x.Artist.Name.Contains(artistname)
+                              orderby x.ReleaseYear, x.Title
+                              select new AlbumsOfArtist
                               {
-                                  AlbumTitle = x.Title,
-                                  Year = x.ReleaseYear,
-                                  ArtistName = x.Artist.Name
+                                  Title = x.Title,
+                                  ArtistName = x.Artist.Name,
+                                  RYear = x.ReleaseYear,
+                                  RLabel = x.ReleaseLabel
                               };
                 return results.ToList();
             }
         }
 
-        #endregion
+        [DataObjectMethod(DataObjectMethodType.Select,false)]
+        public List<AlbumDTO> Album_AlbumAndTracks()
+        {
+            using (var context = new ChinookContext())
+            {
+                var results = from x in context.Albums
+                              where x.Tracks.Count() > 25
+                              select new AlbumDTO
+                              {
+                                  AlbumTitle = x.Title,
+                                  AlbumArtist = x.Artist.Name,
+                                  TrackCount = x.Tracks.Count(),
+                                  PlayTime = x.Tracks.Sum(z => z.Milliseconds),
+                                  AlbumTracks = (from y in x.Tracks
+                                                 select new TrackPOCO
+                                                 {
+                                                     SongName = y.Name,
+                                                     SongGenre = y.Genre.Name,
+                                                     SongLength = y.Milliseconds
+                                                 }).ToList()
 
-        #region Add, Update and Delete
-        [DataObjectMethod(DataObjectMethodType.Insert, false)]
+                              };
+                return results.ToList();
+            }
+        }
+
+        //[DataObjectMethod(DataObjectMethodType.Select, false)]
+        //public List<SelectionList> List_AlbumNames()
+        //{
+        //    using (var context = new ChinookContext())
+        //    {
+        //        var results = from x in context.Albums
+        //                      orderby x.Title
+        //                      select new SelectionList
+        //                      {
+        //                          IDValueField = x.AlbumId,
+        //                          DisplayText = x.Title
+        //                      };
+        //        return results.ToList();
+        //    }
+        //}
+    
+    #endregion
+
+    #region Add, Update, Delete
+    [DataObjectMethod(DataObjectMethodType.Insert,false)]
         public int Album_Add(Album item)
         {
             using (var context = new ChinookContext())
             {
                 if (CheckReleaseYear(item))
                 {
-                    //any additional logic
-                    item.ReleaseLabel = string.IsNullOrEmpty(item.ReleaseLabel) ? null :
-                        item.ReleaseLabel;
-
+                    //item.Title = null; use to test entity validation
                     context.Albums.Add(item);   //staging
-                    context.SaveChanges();      //commit to database
-                    return item.AlbumId;        //return the new identity value of the pkey
+                    context.SaveChanges();      //committed
+                    return item.AlbumId;        //return new id value
                 }
                 else
                 {
@@ -109,19 +138,16 @@ namespace ChinookSystem.BLL
             }
         }
 
-        [DataObjectMethod(DataObjectMethodType.Update, false)]
+        [DataObjectMethod(DataObjectMethodType.Update,false)]
         public int Album_Update(Album item)
         {
             using (var context = new ChinookContext())
             {
                 if (CheckReleaseYear(item))
                 {
-                    //any additional logic
-                    item.ReleaseLabel = string.IsNullOrEmpty(item.ReleaseLabel) ? null :
-                        item.ReleaseLabel;
-
-                    context.Entry(item).State = System.Data.Entity.EntityState.Modified;   //staging
-                    return context.SaveChanges();      //commit to database
+                    context.Entry(item).State =
+                    System.Data.Entity.EntityState.Modified;
+                    return context.SaveChanges();
                 }
                 else
                 {
@@ -135,37 +161,44 @@ namespace ChinookSystem.BLL
         {
             return Album_Delete(item.AlbumId);
         }
-
         public int Album_Delete(int albumid)
         {
             using (var context = new ChinookContext())
             {
-                //physical delete
                 var existing = context.Albums.Find(albumid);
-                context.Albums.Remove(existing);
-                return context.SaveChanges();
+                if (existing == null)
+                {
+                    throw new Exception("Album not on file. Delete unnecessary.");
+                }
+                else
+                {
+                    context.Albums.Remove(existing);
+                    return context.SaveChanges();
+                }
             }
+           
         }
         #endregion
+
         #region Support Methods
         private bool CheckReleaseYear(Album item)
         {
             bool isValid = true;
             int releaseyear;
-            if (string.IsNullOrEmpty(item.ReleaseYear.ToString()))
+            if(string.IsNullOrEmpty(item.ReleaseYear.ToString()))
             {
                 isValid = false;
                 reasons.Add("Release year is required");
             }
-            else if(!int.TryParse(item.ReleaseYear.ToString(),out releaseyear))
+            else if (!int.TryParse(item.ReleaseYear.ToString(),out releaseyear))
             {
                 isValid = false;
-                reasons.Add("Release year is not a numeric year (yyyy)");
+                reasons.Add("Release year is not an number");
             }
-            else if (releaseyear <1950 || releaseyear > DateTime.Today.Year)
+            else if (releaseyear < 1950 || releaseyear > DateTime.Today.Year)
             {
                 isValid = false;
-                reasons.Add(string.Format("Release year of {0} invalid. Year must be between 1950 and today",
+                reasons.Add(string.Format("Album release year of {0} invalid. Year must be between 1950 and today",
                     releaseyear));
             }
             return isValid;
